+++ common/src/main/kotlin/com/cwg/mod/gui/EvolutionsGui.kt
@@ -27,7 +27,7 @@ object EvolutionsGui {
 
     fun open(species: FormData, player: ServerPlayer, page: Int = 0): SimpleGui {
         val gui = SimpleGui(MenuType.GENERIC_9x4, player, false)
-        val redPane = GuiHelper.RED_PANE
+        val redPane = GuiHelper.fillerPane()
 
         gui.title = Component.literal(lang.evolutionsTitle).red()